update float to double

+++ resources.h
@@ -1,8 +1,12 @@
+#include "cmath"
+#include "cfloat"
+#include "limits.h"
+
 // A structure to represent a Point in 2D plane 
 class Point 
 { 
 	public:
-	int x, y; 
+	double x, y; 
 }; 
 
 // Needed to sort array of points 
@@ -14,13 +18,13 @@ int compareY(const void* a, const void* b);
 
 // A utility function to find the 
 // distance between two points 
-float dist(Point p1, Point p2);
+double dist(Point p1, Point p2);
 
 // A Brute Force method to return the 
 // smallest distance between two points 
 // in P[] of size n 
-float bruteForce(Point P[], int n);
+double bruteForce(Point P[], int n);
 
 // A utility function to find 
 // minimum of two float values 
-float min(float x, float y);
+double min(double x, double y);
